refactor(frontend): colocate single-route components
Eight components lived in src/components/ while only one route rendered
each of them. They now sit in that route's _components/ directory:
career-timeline, skill-groups and animated-counter under a-propos,
contact-form under contact, bookmark-composer and bookmark-grid under
veille, availability-badge and tilt under the home route.

What stays in src/components/ is shared on purpose: the site frame and
its client islands, the theme provider, the motion infrastructure, and
project-visual, the only component two routes render.

The files move unchanged; only the import paths follow them.
