+++ apps/frontend/src/app/(site)/(home)/_components/availability-badge.tsx
