+++ apps/frontend/src/app/(site)/a-propos/_components/animated-counter.tsx
