+++ apps/frontend/src/app/(site)/a-propos/_components/skill-groups.tsx
